Parse JSON in the constructor / refactor
