+++ google-cloud-tools-plugin/resources/messages/CloudToolsBundle.properties
@@ -62,12 +62,13 @@ appengine.deployment.error.creating.staging.directory=There was an unexpected er
 appengine.generate.config.button=Generate...
 appengine.appyaml.location.label=app.yaml path\:
 appengine.dockerfile.location.browse.button=Browse to your Dockerfile location
+appengine.deployment.version.extraction.failure=Failed to parse service/version info from the ''gcloud'' deployment output. \nPlease make sure that you are using the latest version of the Google Cloud SDK.\nRun ''gcloud components update'' to update the SDK. (See: https://cloud.google.com/sdk/gcloud/reference/components/update.)
 appengine.deployment.error.with.code=Deployment failed with exit code: {0}\nPlease make sure that you are using the latest version of the Google Cloud SDK.\nRun ''gcloud components update'' to update the SDK. (See: https://cloud.google.com/sdk/gcloud/reference/components/update.)
 appengine.deployment.error.during.execution=Deployment failed due to an unexpected error during the deployment execution phase.
 appengine.stop.modules.version.confirmation.message=<html>Are you sure you want to stop this running version of your project?\nYou can also manage your running versions from the Cloud Platform Console. {0}Learn more{1}.</html>
 appengine.stop.modules.version.confirmation.title=Stop Project Version
-appengine.stop.modules.version.execution.error=Failed to stop application due to an unexpected error during the execution phase.\nPlease make sure that you are using the latest version of the Google Cloud SDK.\nRun ''gcloud components update'' to update the SDK. (See: https://cloud.google.com/sdk/gcloud/reference/components/update.)
-appengine.stop.modules.version.execution.error.with.code=Stop application failed with exit code: {0}\nPlease make sure that you are using the latest version of the Google Cloud SDK.\nRun ''gcloud components update'' to update the SDK. (See: https://cloud.google.com/sdk/gcloud/reference/components/update.)
+appengine.stop.modules.version.execution.error=Failed to stop application due to an unexpected error during the execution phase.
+appengine.stop.modules.version.execution.error.with.code=Stop application failed with exit code: {0}
 appengine.stop.modules.version.success.dialog.message=Application has been successfully stopped.
 appengine.stop.modules.version.success.dialog.title=Application Stopped
 appengine.stop.modules.version.canceled.message=Undeploy Canceled
